# apps/web/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "louped",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Run",
        "operationId": "run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/samples": {
      "get": {
        "summary": "Samples",
        "operationId": "samples_api_runs__run_id__samples_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SampleSummary"
                  },
                  "title": "Response Samples Api Runs  Run Id  Samples Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/samples/{sample_id}": {
      "get": {
        "summary": "Sample",
        "operationId": "sample_api_runs__run_id__samples__sample_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          },
          {
            "name": "epoch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Epoch"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SampleDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/artifacts/{path}": {
      "get": {
        "summary": "Artifact",
        "operationId": "artifact_api_runs__run_id__artifacts__path__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/views": {
      "get": {
        "summary": "Views",
        "operationId": "views_api_runs__run_id__views_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RunView"
                  },
                  "title": "Response Views Api Runs  Run Id  Views Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/features": {
      "get": {
        "summary": "Feature List",
        "operationId": "feature_list_api_runs__run_id__features_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  },
                  "title": "Response Feature List Api Runs  Run Id  Features Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/features/{feature}": {
      "get": {
        "summary": "Feature",
        "operationId": "feature_api_runs__run_id__features__feature__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "feature",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Feature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureDashboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/labels": {
      "get": {
        "summary": "Labels",
        "operationId": "labels_api_runs__run_id__labels_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "enum": ["a", "b", "tie"],
                    "type": "string"
                  },
                  "title": "Response Labels Api Runs  Run Id  Labels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/labels/{sample_id}": {
      "post": {
        "summary": "Label",
        "operationId": "label_api_runs__run_id__labels__sample_id__post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "sample_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sample Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LabelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "enum": ["a", "b", "tie"],
                    "type": "string"
                  },
                  "title": "Response Label Api Runs  Run Id  Labels  Sample Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/agreement": {
      "get": {
        "summary": "Agreement",
        "operationId": "agreement_api_runs__run_id__agreement_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agreement"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/compare": {
      "get": {
        "summary": "Compare",
        "operationId": "compare_api_compare_get",
        "parameters": [
          {
            "name": "a",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "A"
            }
          },
          {
            "name": "b",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "B"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comparison"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/experiments/{name}": {
      "get": {
        "summary": "Experiment",
        "operationId": "experiment_api_experiments__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Agreement": {
        "properties": {
          "labelled": {
            "type": "integer",
            "title": "Labelled",
            "description": "Pairs the person labelled."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Pairs the judge judged."
          },
          "agreement": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agreement",
            "description": "Share of labelled pairs with the same pick."
          },
          "kappa": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kappa",
            "description": "Cohen's kappa: agreement beyond chance."
          }
        },
        "type": "object",
        "required": ["labelled", "total", "agreement", "kappa"],
        "title": "Agreement",
        "description": "A judge run's agreement with a person's labels on its pairs."
      },
      "Artifact": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          }
        },
        "type": "object",
        "required": ["path", "size"],
        "title": "Artifact"
      },
      "Comparison": {
        "properties": {
          "a": {
            "type": "string",
            "title": "A"
          },
          "b": {
            "type": "string",
            "title": "B"
          },
          "only_a": {
            "type": "integer",
            "title": "Only A"
          },
          "only_b": {
            "type": "integer",
            "title": "Only B"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/PairedScore"
            },
            "type": "array",
            "title": "Scores"
          }
        },
        "type": "object",
        "required": ["a", "b", "only_a", "only_b", "scores"],
        "title": "Comparison"
      },
...
      "FeatureDashboard": {
        "properties": {
          "feature": {
            "type": "integer",
            "title": "Feature"
          },
          "hook": {
            "type": "string",
            "title": "Hook"
          },
          "layer": {
            "type": "integer",
            "title": "Layer"
          },
          "density": {
            "type": "number",
            "title": "Density"
          },
          "max": {
            "type": "number",
            "title": "Max"
          },
          "histogram": {
            "$ref": "#/components/schemas/Histogram"
          },
          "promoted": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Promoted"
          },
          "suppressed": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Suppressed"
          },
          "examples": {
            "$ref": "#/components/schemas/TokensView"
          },
          "neuronpedia": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neuronpedia"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "sae": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sae"
          },
          "sae_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sae Id"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "hook",
          "layer",
          "density",
          "max",
          "histogram",
          "promoted",
          "suppressed",
          "examples"
        ],
        "title": "FeatureDashboard",
        "description": "One SAE feature over a dataset, as `louped features` logs it under features/."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LabelRequest": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "enum": ["a", "b", "tie"]
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": ["label"],
        "title": "LabelRequest"
      },
...
      "MetricPoint": {
        "properties": {
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": ["step", "value"],
        "title": "MetricPoint"
      },
...
      "RunDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": ["eval", "analysis", "training"],
            "title": "Kind"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "experiment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiment"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Metrics"
          },
          "samples": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Samples"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Params"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags"
          },
          "history": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/MetricPoint"
              },
              "type": "array"
            },
            "type": "object",
            "title": "History"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/Artifact"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "scorers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scorers"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "log": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "name",
          "experiment",
          "status",
          "created",
          "model",
          "metrics",
          "samples",
          "params",
          "tags",
          "history",
          "artifacts",
          "scorers",
          "error"
        ],
        "title": "RunDetail"
      },
...
      "RunView": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "view": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/HeatmapView"
              },
              {
                "$ref": "#/components/schemas/LineView"
              },
              {
                "$ref": "#/components/schemas/ScatterView"
              },
              {
                "$ref": "#/components/schemas/TableView"
              },
              {
                "$ref": "#/components/schemas/TokensView"
              }
            ],
            "title": "View",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "heatmap": "#/components/schemas/HeatmapView",
                "line": "#/components/schemas/LineView",
                "scatter": "#/components/schemas/ScatterView",
                "table": "#/components/schemas/TableView",
                "tokens": "#/components/schemas/TokensView"
              }
            }
          }
        },
        "type": "object",
        "required": ["path", "view"],
        "title": "RunView",
        "description": "A figure a run logged under views/, addressed by its artifact path."
      },
      "SampleDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "epoch": {
            "type": "integer",
            "title": "Epoch"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/Score"
            },
            "type": "array",
            "title": "Scores"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Metadata"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": ["id", "epoch", "target", "messages", "scores", "metadata", "error"],
        "title": "SampleDetail"
      },
      "SampleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "epoch": {
            "type": "integer",
            "title": "Epoch"
          },
          "input": {
            "type": "string",
            "title": "Input"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "scores": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Scores"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": ["id", "epoch", "input", "target", "scores", "error"],
        "title": "SampleSummary"
      },
...
    }
  }
}